user

+++ app/src/main/java/com/example/project3android/API/PostAPI.java
@@ -7,8 +7,11 @@ import com.example.project3android.Feed.Post.PostDao;
 import com.example.project3android.MyApplication;
 import com.example.project3android.R;
 import com.example.project3android.Repositories.PostRepository;
+import com.example.project3android.User.CurrentUser;
 
 import java.util.List;
+import java.util.concurrent.Executor;
+import java.util.concurrent.Executors;
 
 import retrofit2.Call;
 import retrofit2.Callback;
@@ -27,17 +30,20 @@ public class PostAPI {
         this.dao = dao;
 
         retrofit = new Retrofit.Builder().baseUrl(MyApplication.context.getString(R.string.BaseUrl))
+                .callbackExecutor(Executors.newSingleThreadExecutor())
                 .addConverterFactory(GsonConverterFactory.create()).build();
         webServiceAPI = retrofit.create(WebServiceAPI.class);
     }
     public void get() {
-        Call<List<Post>> call = webServiceAPI.getPosts();
+        Call<List<Post>> call = webServiceAPI.getPosts(CurrentUser.getInstance().getJwtToken());
         call.enqueue(new Callback<List<Post>>() {
         @Override
         public void onResponse(Call<List<Post>> call, Response<List<Post>> response) {
-                postListData.setValue(response.body());
-                // postListData.postValue(response.body());
-                }
+            new Thread(() -> {
+                dao.insert(response.body());
+                postListData.postValue(dao.index());
+            }).start();
+        }
 
         @Override
         public void onFailure(Call<List<Post>> call, Throwable t) {}
@@ -45,10 +51,50 @@ public class PostAPI {
     }
 
     public void add(Post post) {
-        dao.insert(post);
+        // dao.insert(post);
+        Call<Void> call = webServiceAPI.createPost(post, CurrentUser.getInstance().getJwtToken());
+        call.enqueue(new Callback<Void>() {
+            @Override
+            public void onResponse(Call<Void> call, Response<Void> response) {
+                if (response.isSuccessful()) {
+                    new Thread(() -> {
+                        dao.insert(post);
+                        postListData.postValue(dao.index());
+                    }).start();
+                } else {
+                    // Handle unsuccessful response
+                }
+            }
+
+            @Override
+            public void onFailure(Call<Void> call, Throwable t) {
+                // Handle failure
+            }
+        });
     }
 
     public void delete(Post post) {
-        dao.delete(post);
+        //dao.delete(post);
+        Call<Void> call = webServiceAPI.deletePost(
+                CurrentUser.getInstance().getCurrentUser().getId(), post.getId(),
+                CurrentUser.getInstance().getJwtToken());
+        call.enqueue(new Callback<Void>() {
+            @Override
+            public void onResponse(Call<Void> call, Response<Void> response) {
+                if (response.isSuccessful()) {
+                    new Thread(() -> {
+                        dao.delete(post);
+                        postListData.postValue(dao.index());
+                    }).start();
+                } else {
+                    // Handle unsuccessful response
+                }
+            }
+
+            @Override
+            public void onFailure(Call<Void> call, Throwable t) {
+                // Handle failure
+            }
+        });
     }
 }
